modified AsteroidsGame.pde

+++ AsteroidsGame.pde
@@ -31,10 +31,12 @@ public void draw()
     strokeWeight(15);
     rock.get(j).show();
     rock.get(j).move();
-    int d = (int)dist(rock.get(j).getX(), rock.get(j).getY(), flyer.getX(), flyer.getY());
+    /*
+    float d = dist(rock.get(j).getX(), rock.get(j).getY(), flyer.getX(), flyer.getY());
     if (d < 32) {
       rock.remove(j);
     }
+    */
   }
   int size = rock.size();
   if (size < 15) {
